requirements.txt
# Core
numpy
pandas
PyYAML
tqdm

# Image processing
Pillow
opencv-python

# ML + embeddings
torch
torchvision
transformers
scikit-learn

# Jupyter (for development)
notebook